Thrift connections managed by scamandrios now have a lifespan.
At the end of the lifespan, they are torn down & replaced. This should
be transparent to the user of a scamandrios client. The lifespan defaults
to 20 minutes, but you may override this by setting a `TTL` field
on the connection & connection pool options objects.

Connection pools randomly alter the lifespan of each of their
connections so they do not all time out at once. This jitter is set
to 10 seconds right now. We might need to widen this.

There is also a small set of tests for this feature.

Bumped the version to 0.2.0.