Enable connection monitoring and more log events.
Every ten seconds, we do a query of the system keyspaces for all
clients we expect to be alive. If they're not alive (timeout, error),
they go in the dead pool, which we periodically check as well.
We also set their ready flag to false, so we do not attempt to use
those connections in the pool.

We *did* have a way to notice dead nodes at startup, but there was
nothing that noticed dying nodes to take them out of the pool. Now
there is.